requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
requests==2.32.3
python-dotenv==1.0.1
openai==1.54.4
httpx>=0.27,<0.28
